+++ pc_client/PCDiscoveryService.h
@@ -14,7 +14,7 @@ public:
 	PCDiscoveryService();
 	virtual ~PCDiscoveryService();
 
-	virtual uint32_t Discover(std::string clientIP, uint16_t clientDiscoveryPort, std::string serverIP, uint16_t serverDiscoveryPort, ENetAddress& remote) override;
+	virtual uint64_t Discover(std::string clientIP, uint16_t clientDiscoveryPort, std::string serverIP, uint16_t serverDiscoveryPort, ENetAddress& remote) override;
 protected:
 	ENetSocket CreateDiscoverySocket(std::string ip, uint16_t discoveryPort);
 	bool awaiting = false;
